[#250] refactor: 공통 패키지 구조 이동
