+++ linkmind/src/main/java/com/app/toaster/category/controller/CategoryController.java
@@ -4,12 +4,12 @@ import com.app.toaster.category.controller.request.ChangeCateoryPriorityDto;
 import com.app.toaster.category.controller.request.ChangeCateoryTitleDto;
 import com.app.toaster.category.controller.request.CreateCategoryDto;
 import com.app.toaster.common.dto.ApiResponse;
-import com.app.toaster.config.UserId;
+import com.app.toaster.common.config.UserId;
 import com.app.toaster.category.controller.response.CategoriesResponse;
 import com.app.toaster.toast.controller.response.ToastFilter;
 import com.app.toaster.category.controller.response.GetCategoryResponseDto;
-import com.app.toaster.controller.valid.Severity;
-import com.app.toaster.controller.valid.TitleValid;
+import com.app.toaster.utils.valid.Severity;
+import com.app.toaster.utils.valid.TitleValid;
 import com.app.toaster.exception.Success;
 import com.app.toaster.category.service.CategoryService;
 import com.app.toaster.search.service.SearchService;
