+++ linkmind/src/main/java/com/app/toaster/common/config/UserIdResolver.java
@@ -1,13 +1,12 @@
-package com.app.toaster.config;
+package com.app.toaster.common.config;
 
+import com.app.toaster.common.config.jwt.JwtService;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-
-import com.app.toaster.config.jwt.JwtService;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.constraints.NotNull;
