--- linkmind/src/main/java/com/app/toaster/controller/valid/marker/ToastValidationGroup.java
@@ -1,4 +1,0 @@
-package com.app.toaster.controller.valid.marker;
-
-public interface ToastValidationGroup {
-}
