+++ linkmind/src/main/java/com/app/toaster/popup/controller/PopupController.java
@@ -3,10 +3,8 @@ package com.app.toaster.popup.controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.app.toaster.popup.controller.request.PopUpRequestDto;
@@ -14,7 +12,7 @@ import com.app.toaster.popup.controller.response.InvisibleResponseDto;
 import com.app.toaster.popup.controller.response.PopupResponseDto;
 import com.app.toaster.popup.service.PopupService;
 import com.app.toaster.common.dto.ApiResponse;
-import com.app.toaster.config.UserId;
+import com.app.toaster.common.config.UserId;
 import com.app.toaster.exception.Success;
 
 import jakarta.validation.Valid;
