+++ linkmind/src/main/java/com/app/toaster/toast/controller/ToastController.java
@@ -20,7 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.app.toaster.common.dto.ApiResponse;
 // import com.app.toaster.config.UserId;
-import com.app.toaster.config.UserId;
+import com.app.toaster.common.config.UserId;
 import com.app.toaster.exception.Success;
 import com.app.toaster.parse.service.ParsingService;
 import com.app.toaster.toast.service.ToastService;
