+++ linkmind/src/main/java/com/app/toaster/toast/controller/request/SaveToastDto.java
@@ -1,9 +1,4 @@
 package com.app.toaster.toast.controller.request;
 
-import java.util.List;
-
-import com.app.toaster.controller.valid.Severity;
-import com.app.toaster.controller.valid.TitleValid;
-
 public record SaveToastDto(String linkUrl, Long categoryId) {
 }
