+++ linkmind/src/main/java/com/app/toaster/utils/health/HealthCheckController.java
@@ -1,4 +1,4 @@
-package com.app.toaster.controller;
+package com.app.toaster.utils.health;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
