+++ linkmind/src/main/java/com/app/toaster/utils/valid/marker/ToastValidationGroup.java
@@ -1,0 +1,4 @@
+package com.app.toaster.utils.valid.marker;
+
+public interface ToastValidationGroup {
+}
